// components/button/buttonTargets.enum.ts
export enum ButtonTargets {
  button = 'button',
  errorsRoot = 'errorsRoot',
  errorsUnorderedList = 'errorsUnorderedList',
  errorsUnorderedListItem = 'errorsUnorderedListItem',
  errorsFieldError = 'errorsFieldError',
  label = 'label',
  leftIcon = 'leftIcon',
  loadingIcon = 'loadingIcon',
  rightIcon = 'rightIcon',
  root = 'root',
  simpleViewRoot = 'simpleViewRoot',
  simpleViewButton = 'simpleViewButton',
}